Move multi files to shuffle/
