+++ distributed/shuffle/tests/test_multi_file.py
@@ -5,17 +5,18 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from distributed.multi_file import MultiFile
+from distributed.shuffle.multi_file import MultiFile
 
 
-def test_basic(tmp_path):
+@pytest.mark.asyncio
+async def test_basic(tmp_path):
     with MultiFile(
         directory=tmp_path, n_files=4, memory_limit="16 MiB", join=pd.concat
     ) as mf:
         df = pd.DataFrame({"x": np.arange(1000), "y": np.arange(1000) * 2})
-        mf.write(df, "a")
-        mf.write(df, "b")
-        mf.write(df * 2, "a")
+        await mf.write(df, "a")
+        await mf.write(df, "b")
+        await mf.write(df * 2, "a")
 
         a = mf.read("a")
         b = mf.read("b")
@@ -26,8 +27,9 @@ def test_basic(tmp_path):
     assert not os.path.exists(tmp_path)
 
 
+@pytest.mark.asyncio
 @pytest.mark.parametrize("count", [2, 100, 1000])
-def test_many(tmp_path, count):
+async def test_many(tmp_path, count):
     with MultiFile(
         directory=tmp_path, n_files=4, memory_limit="16 MiB", join=pd.concat
     ) as mf:
@@ -37,11 +39,11 @@ def test_many(tmp_path, count):
 
         random.shuffle(L)
         for i in L:
-            mf.write(df + i, i)
+            await mf.write(df + i, i)
 
         random.shuffle(L)
         for i in L:
-            mf.write(df * i, i)
+            await mf.write(df * i, i)
 
         random.shuffle(L)
         for i in L:
